feat: Gen Z UI modernization with green-first color palette
- Rewrite color palette: green primary (0xFF16A34A), teal secondary, lime accent
- Modernize login/register screen: remove AppBar, add large headers, 14px radius inputs
- Bold clickable Register link navigating to Create Account screen
- Dashboard overhaul: gradient hero card, color-coded tax tiles with subtitles
- Welcome screen: 3-stop green gradient, contemporary Sign In/Create Account buttons
- AppBar updated to new primaryDark green (0xFF166534)
- Tailwind-inspired Slate neutrals for text and backgrounds
- All builds verified: Web + Android APK

+++ Documents/Trykon/taxng_advisor/lib/features/cit/presentation/cit_calculator_screen.dart
@@ -84,8 +84,8 @@ class _CitCalculatorScreenState extends State<CitCalculatorScreen>
         final profit = data['profit']!;
 
         return CitCalculator.calculate(
-          turnover: turnover as double,
-          profit: profit as double,
+          turnover: turnover,
+          profit: profit,
         );
       },
       operationName: 'CIT Calculation',
